Add synchronization to buffer access and add log size limit
