+++ shared/src/commonMain/kotlin/org/jetbrains/kotlinconf/utils/DebugLogger.kt
@@ -10,6 +10,9 @@ class DebugLogger(private val platformLogger: Logger) : Logger {
 
     override fun log(tag: String, lazyMessage: () -> String) {
         logs += "${Clock.System.now()} [${tag}] ${lazyMessage()}"
+        while (logs.size > MAX_LOG_MESSAGES_IN_MEMORY) {
+            logs.removeAt(0)
+        }
         platformLogger.log(tag, lazyMessage)
     }
